refactor: types improvements
