+++ tsconfig.json
@@ -5,6 +5,6 @@
     "declarationDir": "lib/types/",
     "strict": true
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "test/**/*.ts"],
   "exclude": ["node_modules"]
 }
